refactor: extract shared utilities to cli/util.ts
